requirements.txt
# Workhours PIP requirements file


# collector requirements
#file:///srv/repos/etc/requirements-sqlalchemy.txt
pytz
BeautifulSoup  # 3.2
#BeautifulSoup4 # TODO
pyutmp


# webapp requirements
waitress
pyramid

passlib

formencode
webhelpers
iso8601
colander
peppercorn

deform
deform_jinja2
pyramid_deform

pyramid_simpleform

pyramid_sqlalchemy
fixture

# TODO: src
pyramid_restler

celery
pyramid_celery

# debugging
pyramid_debugtoolbar
